bot/requirements.txt
aiogram==3.7.0
python-dotenv
requests
httpx==0.27.0